Move reload-related functions to reload.go
